requirements.txt
certifi==2019.9.11
Click==7.0
dash==1.4.1
dash-core-components==1.3.1
dash-html-components==1.0.1
dash-renderer==1.1.2
dash-table==4.4.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mysql-connector-python==8.0.18
numpy==1.17.3
pandas==0.25.3
plotly==4.3.0
protobuf==3.10.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
retrying==1.3.3
six==1.13.0
Werkzeug==0.16.0
